refactor directory structures one more
