+++ src/Inspector/Daemon/Reader/PhpReaderTask.php
@@ -11,7 +11,7 @@
 
 declare(strict_types=1);
 
-namespace PhpProfiler\Lib\Concurrency\Amphp\Task;
+namespace PhpProfiler\Inspector\Daemon\Reader;
 
 use Amp\Parallel\Sync\Channel;
 use Generator;
@@ -19,7 +19,6 @@ use PhpProfiler\Inspector\Settings\GetTraceSettings;
 use PhpProfiler\Inspector\Settings\TargetPhpSettings;
 use PhpProfiler\Inspector\Settings\TargetProcessSettings;
 use PhpProfiler\Inspector\Settings\TraceLoopSettings;
-use PhpProfiler\Inspector\Daemon\Worker\ReaderLoopProvider;
 use PhpProfiler\Lib\PhpProcessReader\PhpGlobalsFinder;
 use PhpProfiler\Lib\PhpProcessReader\PhpMemoryReader\ExecutorGlobalsReader;
